// src/client/models/ReportQueryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReportQueryDto = {
    Interval?: string;
    Projects?: Array<string>;
    Tags?: Array<string>;
    Employees?: Array<string>;
    Teams?: Array<string>;
    Billable?: string;
    Group1?: string;
    Group2?: string;
};